Basic template support
This commit adds basic support for EEx templates to Spaceboy. This is
accomplished through a new Spaceboy.Controller module that mimics
Phoenix.Controller.

Templates are loaded from a template root, and mimetypes are inferred
from the file name.

I also updated the example application with templates and the changes
caused by the introduction of Spaceboy.Controller.

Possible Issues:

The template root is calculated using File.cwd! which, depending on the
project type, may or may not point to the right place.
I attempted to figure out how Phoenix does this, but was unable to make
sense of their templating code.

I updated the todo list in the REAME, but I was unsure what "ANSI files"
meant so I left it un-checked. The current templating system should
support all MIME types, so long as it can guess it from the file
extension.

Edit 1:

Merged in master and fixed conflicts, and then ran Credo and fixed lint
issues (it caught a couple good ones).

Edit 2:

Added an example of using ANSI escape codes, and simplified the
behavior of the root path and deleted the comment.

Signed-off-by: Steven vanZyl <rushsteve1@rushsteve1.us>